major backend refactor
